snapshot: colors

+++ magick-kt-native-q8/src/commonMain/kotlin/imagemagick/helpers/PercentageHelper.kt
@@ -15,5 +15,5 @@ object PercentageHelper {
 
     // TODO Check this nightmare
     fun toQuantumType(value: Percentage): QuantumType =
-        QuantumImpl.max.toDouble().times(value.toDouble().div(100)).toFloat().toUInt().toUByte()
+        QuantumImpl.max.toDouble().times(value.toDouble().div(100)).toUInt().toUByte()
 }
